Move the recursive service to its own package
