+++ Go/Recursion/RecursiveFibonacciService.go
@@ -1,4 +1,4 @@
-package main
+package recursion
 
 import (
 	"Fibonacci/common"
